unified a bunch of elements between recommendations and map page
