+++ src/components/showMap.jsx
@@ -7,8 +7,8 @@ import { Header } from '../components/utils/header';
 import { Footer } from '../components/utils/footer';
 import { useParams } from "react-router-dom";
 import { WindSpeedLayer, TemperatureLayer, CloudLayer, RainViewerData, HybridLayer } from './layers';
-import { MapMode } from './utils/mapMode';
 import { WeatherPopupContent } from './utils/weatherVariables';
+import { CustomZoomControl, CustomAttributionControl, MapMode } from './utils/mapElements';
 
 // Custom component to style popup containers based on map mode
 const CustomPopupStyle = ({ mapType }) => {
@@ -93,159 +93,6 @@ const CustomPopupStyle = ({ mapType }) => {
     return () => {
       map.off('popupopen', stylePopups);
     };
-  }, [mapType, map]);
-  
-  return null; // This component doesn't render anything, just applies styling
-};
-
-// Custom component to style zoom control based on map mode
-const CustomZoomControl = ({ mapType }) => {
-  const map = useMap();
-  
-  useEffect(() => {
-    // Find all zoom control elements and style them according to map type
-    const zoomInButton = document.querySelector('.leaflet-control-zoom-in');
-    const zoomOutButton = document.querySelector('.leaflet-control-zoom-out');
-    const zoomContainer = document.querySelector('.leaflet-control-zoom');
-    
-    if (zoomInButton && zoomOutButton && zoomContainer) {
-      // Apply styles to the container
-      zoomContainer.style.border = 'none';
-      zoomContainer.style.boxShadow = mapType === 'light' 
-        ? '0 2px 10px rgba(0, 0, 0, 0.1)' 
-        : '0 2px 10px rgba(0, 0, 0, 0.4)';
-      zoomContainer.style.borderRadius = '8px';
-      zoomContainer.style.overflow = 'hidden';
-      
-      // Common styles for both buttons
-      [zoomInButton, zoomOutButton].forEach(button => {
-        button.style.transition = 'all 0.2s ease';
-        button.style.border = 'none';
-        button.style.width = '36px';
-        button.style.height = '36px';
-        button.style.lineHeight = '36px';
-        button.style.fontSize = '18px';
-        button.style.fontWeight = 'bold';
-        
-        // Hover effect
-        button.onmouseover = () => {
-          if (mapType === 'light') {
-            button.style.backgroundColor = '#f0f0f0';
-          } else {
-            button.style.backgroundColor = '#444';
-          }
-        };
-        
-        // Reset on mouseout
-        button.onmouseout = () => {
-          if (mapType === 'light') {
-            button.style.backgroundColor = '#fff';
-          } else {
-            button.style.backgroundColor = '#333';
-          }
-        };
-      });
-      
-      // Apply theme-specific styles
-      if (mapType === 'light') {
-        // Light mode styling
-        zoomInButton.style.color = '#333';
-        zoomInButton.style.backgroundColor = '#fff';
-        zoomOutButton.style.color = '#333';
-        zoomOutButton.style.backgroundColor = '#fff';
-      } else {
-        // Dark mode styling
-        zoomInButton.style.color = '#f0f0f0';
-        zoomInButton.style.backgroundColor = '#333';
-        zoomOutButton.style.color = '#f0f0f0';
-        zoomOutButton.style.backgroundColor = '#333';
-      }
-    }
-  }, [mapType, map]);
-  
-  return null; // This component doesn't render anything, just applies styling
-};
-
-// Custom component to style attribution control based on map mode
-const CustomAttributionControl = ({ mapType }) => {
-  const map = useMap();
-  
-  useEffect(() => {
-    // Find attribution control element and style according to map type
-    const attributionElement = document.querySelector('.leaflet-control-attribution');
-    
-    if (attributionElement) {
-      // Apply common styles
-      attributionElement.style.transition = 'all 0.2s ease';
-      attributionElement.style.padding = '5px 8px';
-      attributionElement.style.fontSize = '11px';
-      attributionElement.style.lineHeight = '1.2';
-      attributionElement.style.boxShadow = mapType === 'light' 
-        ? '0 1px 5px rgba(0, 0, 0, 0.1)' 
-        : '0 1px 5px rgba(0, 0, 0, 0.4)';
-      attributionElement.style.whiteSpace = 'nowrap';
-      attributionElement.style.overflow = 'hidden';
-      attributionElement.style.textOverflow = 'ellipsis';
-      attributionElement.style.maxWidth = 'calc(100vw - 20px)';
-      
-      // Add hover behavior to show full text
-      attributionElement.onmouseover = () => {
-        attributionElement.style.whiteSpace = 'normal';
-        attributionElement.style.maxWidth = '300px';
-        attributionElement.style.zIndex = '1000';
-      };
-      
-      attributionElement.onmouseout = () => {
-        attributionElement.style.whiteSpace = 'nowrap';
-        attributionElement.style.overflow = 'hidden';
-        attributionElement.style.textOverflow = 'ellipsis';
-        attributionElement.style.maxWidth = 'calc(100vw - 20px)';
-      };
-      
-      // Improve links styling
-      const links = attributionElement.querySelectorAll('a');
-      links.forEach(link => {
-        link.style.textDecoration = 'none';
-        link.style.fontWeight = '500';
-        link.style.transition = 'color 0.2s ease';
-        
-        if (mapType === 'light') {
-          link.style.color = '#0078A8';
-        } else {
-          link.style.color = '#6BB0FF';
-        }
-        
-        // Add hover effect
-        link.onmouseover = () => {
-          if (mapType === 'light') {
-            link.style.color = '#005580';
-          } else {
-            link.style.color = '#99CCFF';
-          }
-        };
-        
-        link.onmouseout = () => {
-          if (mapType === 'light') {
-            link.style.color = '#0078A8';
-          } else {
-            link.style.color = '#6BB0FF';
-          }
-        };
-      });
-      
-      // Apply theme-specific styles
-      if (mapType === 'light') {
-        // Light mode
-        attributionElement.style.color = '#333';
-        attributionElement.style.backgroundColor = 'rgba(255, 255, 255, 0.9)';
-        attributionElement.style.borderRadius = '8px 0 0 0';
-      } else {
-        // Dark mode
-        attributionElement.style.color = '#e0e0e0';
-        attributionElement.style.backgroundColor = 'rgba(40, 40, 40, 0.9)';
-        attributionElement.style.borderRadius = '8px 0 0 0';
-      }
-    }
   }, [mapType, map]);
   
   return null; // This component doesn't render anything, just applies styling
